+++ integration-tests/ssr/pages-using-engines.js
@@ -1,0 +1,5 @@
+const path = require(`path`)
+
+module.exports = {
+  "/ssr1/": path.join(__dirname, `prod-engines-output`, `ssr1.html`),
+}