+++ packages/gatsby-plugin-utils/src/polyfill-remote-file/graphql/gatsby-image-resolver.ts
@@ -1,5 +1,5 @@
-import path from "path"
-import { generatePublicUrl, generateImageArgs } from "../utils/url-generator"
+import { hasFeature } from "../../has-feature"
+import { generateImageUrl } from "../utils/url-generator"
 import { getImageFormatFromMimeType } from "../utils/mime-type-helpers"
 import { stripIndent } from "../utils/strip-indent"
 import {
@@ -7,10 +7,10 @@ import {
   shouldDispatch,
 } from "../jobs/dispatchers"
 import { generatePlaceholder, PlaceholderType } from "../placeholder-handler"
-import { ImageCropFocus, ImageFit, isImage } from "../types"
+import { ImageCropFocus, isImage } from "../types"
 import { validateAndNormalizeFormats, calculateImageDimensions } from "./utils"
 
-import type { Actions } from "gatsby"
+import type { Actions, Store } from "gatsby"
 import type {
   IRemoteFileNode,
   IRemoteImageNode,
@@ -72,10 +72,15 @@ const DEFAULT_PIXEL_DENSITIES = [0.25, 0.5, 1, 2]
 const DEFAULT_BREAKPOINTS = [750, 1080, 1366, 1920]
 const DEFAULT_QUALITY = 75
 
+const GATSBY_SHOULD_TRACK_IMAGE_CDN_URLS = [`true`, `1`].includes(
+  process.env.GATSBY_SHOULD_TRACK_IMAGE_CDN_URLS || ``
+)
+
 export async function gatsbyImageResolver(
   source: IRemoteFileNode,
   args: IGatsbyImageDataArgs,
-  actions: Actions
+  actions: Actions,
+  store?: Store
 ): Promise<{
   images: IGatsbyImageData
   layout: string
@@ -158,10 +163,16 @@ export async function gatsbyImageResolver(
 
     return 1
   }
+
   const sortedFormats = Array.from(formats).sort(
     (a, b) => getFormatValue(b) - getFormatValue(a)
   )
 
+  // Result will be used like this
+  // <picture>
+  // for each result.sources we create a <source srcset="..." /> tag
+  // <img src="fallbacksrc" srcset="fallbacksrcset" />
+  // </picture>
   for (const format of sortedFormats) {
     let fallbackSrc: string | undefined = undefined
     const images = imageSizes.sizes.map(width => {
@@ -169,31 +180,29 @@ export async function gatsbyImageResolver(
         dispatchLocalImageServiceJob(
           {
             url: source.url,
-            extension: format,
-            basename: path.basename(
-              source.filename,
-              path.extname(source.filename)
-            ),
+            mimeType: source.mimeType,
+            filename: source.filename,
+            contentDigest: source.internal.contentDigest,
+          },
+          {
             width,
             height: Math.round(width / imageSizes.aspectRatio),
             format,
-            fit: args.fit as ImageFit,
-            contentDigest: source.internal.contentDigest,
+            cropFocus: args.cropFocus,
             quality: args.quality as number,
           },
-          actions
+          actions,
+          store
         )
       }
 
-      const src = `${generatePublicUrl(source)}/${generateImageArgs({
+      const src = generateImageUrl(source, {
         width,
         height: Math.round(width / imageSizes.aspectRatio),
         format,
         cropFocus: args.cropFocus,
         quality: args.quality as number,
-      })}/${encodeURIComponent(
-        path.basename(source.filename, path.extname(source.filename))
-      )}.${format}`
+      })
 
       if (!fallbackSrc) {
         fallbackSrc = src
@@ -208,7 +217,8 @@ export async function gatsbyImageResolver(
       }
     })
 
-    if (format === sourceMetadata.format && fallbackSrc) {
+    // The latest format (by default will be jpg/png) is the fallback and doesn't need sources
+    if (format === sortedFormats[sortedFormats.length - 1] && fallbackSrc) {
       result.fallback = {
         src: fallbackSrc,
         srcSet: createSrcSetFromImages(images),
@@ -227,7 +237,8 @@ export async function gatsbyImageResolver(
   if (args.placeholder !== `none`) {
     const { fallback, backgroundColor: bgColor } = await generatePlaceholder(
       source,
-      args.placeholder as PlaceholderType
+      args.placeholder as PlaceholderType,
+      store
     )
 
     if (fallback) {
@@ -236,6 +247,11 @@ export async function gatsbyImageResolver(
     if (bgColor) {
       backgroundColor = bgColor
     }
+  }
+
+  // Check if addGatsbyImageSourceUrl for backwards compatibility with older Gatsby versions
+  if (GATSBY_SHOULD_TRACK_IMAGE_CDN_URLS && actions.addGatsbyImageSourceUrl) {
+    actions.addGatsbyImageSourceUrl(source.url)
   }
 
   return {
@@ -250,14 +266,15 @@ export async function gatsbyImageResolver(
 
 export function generateGatsbyImageFieldConfig(
   enums: ReturnType<typeof getRemoteFileEnums>,
-  actions: Actions
+  actions: Actions,
+  store?: Store
 ): IGraphQLFieldConfigDefinition<
   IRemoteFileNode | IRemoteImageNode,
   ReturnType<typeof gatsbyImageResolver>,
   IGatsbyImageDataArgs
 > {
   return {
-    type: `JSON`,
+    type: hasFeature(`graphql-typegen`) ? `GatsbyImageData` : `JSON`,
     description: `Data used in the <GatsbyImage /> component. See https://gatsby.dev/img for more info.`,
     args: {
       layout: {
@@ -356,7 +373,7 @@ export function generateGatsbyImageFieldConfig(
       },
     },
     resolve(source, args): ReturnType<typeof gatsbyImageResolver> {
-      return gatsbyImageResolver(source, args, actions)
+      return gatsbyImageResolver(source, args, actions, store)
     },
   }
 }