+++ packages/gatsby-source-contentful/src/extend-node-type.js
@@ -38,7 +38,6 @@ export async function setFieldsOnGraphQLNodeType({ type, cache }) {
         },
         quality: {
           type: GraphQLInt,
-          defaultValue: 50,
         },
       }
     )