+++ packages/gatsby/src/bootstrap/__mocks__/get-config/compiled-dir/compiled/gatsby-config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  siteMetadata: {
+    title: `compiled`,
+    siteUrl: `https://www.yourdomain.tld`,
+  },
+  plugins: [],
+}