+++ packages/gatsby-source-contentful/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.14.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.14.0/packages/gatsby-source-contentful) (2022-06-07)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.16)
+
+#### Features
+
+- Add setRequestHeaders action/api [#35655](https://github.com/gatsbyjs/gatsby/issues/35655) ([f520e59](https://github.com/gatsbyjs/gatsby/commit/f520e59599a237b0a8ed105bff931c5058e2bfa8))
+- new useContentfulHook to create images on client side [#29263](https://github.com/gatsbyjs/gatsby/issues/29263) ([8d8a800](https://github.com/gatsbyjs/gatsby/commit/8d8a800a40da84c839b1ad73a25648abbc736250))
+
+#### Bug Fixes
+
+- Update `createRemoteFileNode` args [#35422](https://github.com/gatsbyjs/gatsby/issues/35422) fixes [#35363](https://github.com/gatsbyjs/gatsby/issues/35363) ([caaf7a5](https://github.com/gatsbyjs/gatsby/commit/caaf7a555bb42043044f05743eec0ac1ad372cc4))
+
+#### Other Changes
+
+- respect gatsby-plugin-image defaults and extend tests [#33536](https://github.com/gatsbyjs/gatsby/issues/33536) ([2fc7732](https://github.com/gatsbyjs/gatsby/commit/2fc773223bcc5154332b8c0d28f728654998d57d))
+
+### [7.13.1](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.13.1/packages/gatsby-source-contentful) (2022-06-01)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+## [7.13.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.13.0/packages/gatsby-source-contentful) (2022-05-24)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.15)
+
+#### Other Changes
+
+- Add RemoteFile interface [#35617](https://github.com/gatsbyjs/gatsby/issues/35617) ([55a4550](https://github.com/gatsbyjs/gatsby/commit/55a4550d7ba70a5b549a58484496113f27a8315a))
+
+## [7.12.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.12.0/packages/gatsby-source-contentful) (2022-05-10)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.14)
+
+#### Features
+
+- add support for tables in Rich Text [#33870](https://github.com/gatsbyjs/gatsby/issues/33870) ([55c7183](https://github.com/gatsbyjs/gatsby/commit/55c718315f3e60e0448af8c3ef5c19fa8d9384b5))
+- Add `contentTypeFilter` [#35204](https://github.com/gatsbyjs/gatsby/issues/35204) ([7477b5f](https://github.com/gatsbyjs/gatsby/commit/7477b5f36e841f4b60ef8d21e73a01d5251dbf1d))
+
+#### Bug Fixes
+
+- update dependency fs-extra to ^10.1.0 [#34976](https://github.com/gatsbyjs/gatsby/issues/34976) ([10752ed](https://github.com/gatsbyjs/gatsby/commit/10752ed325ac0ebc2655b994862f66abe072e09f))
+- update dependency semver to ^7.3.7 [#35543](https://github.com/gatsbyjs/gatsby/issues/35543) ([f593e00](https://github.com/gatsbyjs/gatsby/commit/f593e005e4561b423a297e967205d833451a2f55))
+- Prevent null pointer exception [#35244](https://github.com/gatsbyjs/gatsby/issues/35244) ([7bc7cf6](https://github.com/gatsbyjs/gatsby/commit/7bc7cf665f07e722d77628d6eabed9c11840e29a))
+
+## [7.11.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.11.0/packages/gatsby-source-contentful) (2022-04-26)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.13)
+
+#### Bug Fixes
+
+- Prevent null pointer exception [#35244](https://github.com/gatsbyjs/gatsby/issues/35244) [#35492](https://github.com/gatsbyjs/gatsby/issues/35492) ([dc25387](https://github.com/gatsbyjs/gatsby/commit/dc25387f4e23ea13cd48817420c733085848807d))
+
+### [7.10.1](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.10.1/packages/gatsby-source-contentful) (2022-04-13)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+## [7.10.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.10.0/packages/gatsby-source-contentful) (2022-04-12)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.12)
+
+#### Bug Fixes
+
+- handle backreferences on data updates properly [#35214](https://github.com/gatsbyjs/gatsby/issues/35214) ([cf98027](https://github.com/gatsbyjs/gatsby/commit/cf980279eabc6ad5b4a001666b69f7497c21bed1))
+
+### [7.9.1](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.9.1/packages/gatsby-source-contentful) (2022-03-31)
+
+#### Bug Fixes
+
+- handle backreferences on data updates properly [#35214](https://github.com/gatsbyjs/gatsby/issues/35214) [#35269](https://github.com/gatsbyjs/gatsby/issues/35269) ([b5f0197](https://github.com/gatsbyjs/gatsby/commit/b5f01971d7572edf6733876251fca4c4f6fe4a36))
+
 ## [7.9.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-source-contentful@7.9.0/packages/gatsby-source-contentful) (2022-03-29)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.11)